Verify if logfile exists before trying to remove it

+++ TraktForVLC.py
@@ -34,7 +34,8 @@ class TraktForVLC(object):
     if LOG_LEVEL is logging.DEBUG:
       logfile = datadir + "/logs/TraktForVLC-DEBUG.log"
       # Remove existing DEBUG file
-      os.remove(logfile)
+      if os.path.isfile(logfile):
+        os.remove(logfile)
     else:
       logfile = datadir + "/logs/TraktForVLC-" + DATETIME.strftime("%y-%m-%d-%H-%M") + ".log"
